+++ src/include/vkc/descriptor/binding.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 #include <utility>
 
 #include <vulkan/vulkan.hpp>
@@ -13,7 +14,7 @@ template <CSupportDraftDescSetLayoutBinding... TManager>
 [[nodiscard]] static constexpr inline auto genDescSetLayoutBindings(const TManager&... mgrs) {
     const auto genDescSetLayoutBinding = [](const auto& mgr, const size_t index) {
         vk::DescriptorSetLayoutBinding binding = mgr.draftDescSetLayoutBinding();
-        binding.setBinding(index);
+        binding.setBinding((uint32_t)index);
         return binding;
     };
