+++ src/include/vkc/descriptor/set.hpp
@@ -2,6 +2,7 @@
 
 #include <array>
 #include <cstddef>
+#include <cstdint>
 #include <memory>
 #include <span>
 #include <utility>
@@ -19,7 +20,7 @@ template <CSupportDraftWriteDescSet... TManager>
 [[nodiscard]] static constexpr inline auto genWriteDescSets(const TManager&... mgrs) {
     const auto genWriteDescSet = [](const auto& mgr, const size_t index) {
         vk::WriteDescriptorSet writeDescSet = mgr.draftWriteDescSet();
-        writeDescSet.setDstBinding(index);
+        writeDescSet.setDstBinding((uint32_t)index);
         return writeDescSet;
     };
